refactor(cplusplus): full refactor and improvement of the interpreter
removed the vector class, because it was corrupting the code, more specifically the stack. added a TuringMachine class, to encapsulate the variables needed and the functions. improved the source file handling and reading, now its more reliable and fast. improve the error handling, now it has a description, a custom code and it is shown with std::cerr, instead of std::cout (todo: change the other error messages to use std::cerr). with the addition of the TuringMachine class, the attributes names were changed to something more official and understandable, with also two methods, one for the generation of the code/instructions and the other for the execution.